+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes for Desonity package will be documented in this file
 
+## 2.0.0 - 2033-03-31
+
+### Changed
+
+- **Massive Changes**, Replaced `IEnumerator`s with `async/await`s.
+- `Route.POST(...)` now returns a JSON string of the format `{"Response":data}` if the post request was successful or `{"Error":error}` if it was not successful.
+
+### Added
+
+- Setup `Identity` so users can create an identity object for whatever public key they want and pass this object to other classes like `Profile` and `Nft` to use their functions.
+- Users can now make posts to DeSo using `Profile.createPost(...)`.
+- Signing of transactions will be done through the Flask Backend until someone figures out how to do the signing on the Unity side in C# (help me pwease :'|)
+
+### TODO
+
+- Replace current string JSON response of `Route.POST(...)` with a more easily useable Object thingy idk.
+- Add code for `READ_WRITE_IDENTITY` scope to use Identity window for signing transactions on the users side.
+
 ## 1.0.2 - 2022-03-27
 
 ### Changed